[CE-121] [CE-122] [CE-123] vSphere agent driver code.
Driver code for create Host and Cluster
on vSphere.
To user they will only see 1 VC host in the
front. the driver will responsible for manage the
lifecycle of the VMs.

In this release only 1 vm will created, in later release
will support create mutiple vms with in 1 VC. those vms are
transparent to User.

The cluster operation on vSphere node will use the
docker-compose directly.

Change-Id: I3c427d998c4ad88ba86e734442c08683a10186b9
Signed-off-by: hainingzhang <haininghenryzh@vmware.com>