[CE-178] Allow OpenStack public IP not to be assigned
Currently the OpenStack provision module always assumes that
the public IP is always needed to access VMs. However, in
some cases, the public IP is not needed and preferred not
to use the resources. This patch set added a new flag named
assign_public_ip to control this behavior. When the flat
is set to true, then public IP address will be assigned to
the virtual machines, if the flag is set to false, then
the IP address won't be assigned.

Change-Id: I8a39571c41960b2c6644e0b93d7bbe2ea4f1d36e
Signed-off-by: tongliofcary <litong01@us.ibm.com>