[CE-255] Allow ansible agent to avoid build any thing
Currently ansible agent requires download fabric source code to
at least build cryptogen and configtxgen binary to produce
org certificats. With this change, agent now can simply download
binaries and docker images from a configuration location such as
a location docker hub. This patch set also improves the docker
images download so that the not needed images will not be pulled
off of the hub. It also contains the recent changes on fabric
for peers to have a listen address and port to be configured.

Change-Id: Ic132fe8bc3ece96e6b99990664eec21b0b9cd568
Signed-off-by: tongliofcary <litong01@us.ibm.com>