[CE-257] Fix chain starting health checking bug
The cluster may fail due to the health check and hang to that status.

This patchset introduces a new 'status' field to track the cluter status
and also updates the agent artifacts.

The fabric v0.6 and syslog options are removed in case others may
choose. Will deprecate those related code in following patchsets.

Change-Id: Ib17acc97411bcf646aae9b27a47a2479fdd97727
Signed-off-by: Baohua Yang <yangbaohua@gmail.com>