[CE-497] Extend the connection profiles
When a fabric network has nodes from multiple cloud, the
endpoints must be using public DNS names or IP addresses.
A section named entityMatchers is needed in the
connection profile for clients to be able to connect to
the network. We need to extend the connection profile
to include that section.

Change-Id: Ie22d758f196430f08b85f2628beefbac87304413
Signed-off-by: Tong Li <litong01@us.ibm.com>