[#issue-319] There is a risk of obtaining expired information
If the file has already been created. Then an exception occurs, at which point the file is not deleted, and the next time the function is called, the information continues to be written to the file, causing the user to get an exception.

Perhaps using files is not a good method;

Close #issue-319

Signed-off-by: zhaoshihong <523713078@qq.com>
Signed-off-by: tianxuanhong <523713078@qq.com>