Nominate Xichen Pan as maintainer
Xichen Pan was the intern student last year, he is actively contributing
into the project in terms of code, features, and bug fixes.

After his intern program, he continues to contribute to the project as
an individual developer for more than half a year.

I believe the nomination will help the project move on more efficiently.

Signed-off-by: Baohua Yang <yangbaohua@gmail.com>